feat: process management + os specific settings + proxied events + launch half done

+++ packages/entity/src/entities/cluster_groups.rs
@@ -8,7 +8,7 @@ use serde::{Deserialize, Serialize};
 #[onelauncher_macro::specta]
 pub struct Model {
 	#[sea_orm(primary_key)]
-	pub id: i32,
+	pub id: u64,
 	#[sea_orm(column_type = "Text")]
 	pub name: String,
 }
